conn-requests: emit NewChannels and ChannelClosed signals.
If the channel manager is a GabbleChannelManager, we assume it will emit
channel-closed. If it's just a TpChannelFactoryIface, we act as though it had
emitted channel-closed in response to each closed signal - this causes the
signal ordering to be wrong in the case of a respawning channel, so factories
for respawning channels will have to implement GabbleChannelManager.

20080725124234-53eee-d65e8cba944f2f20a8ca0b0d1da90406475d29ca.gz