generate MUC presence messages using the real connection-wide presence
With a pedantic interpretation, this is arguable a spec violation because it
exposes your presence to handles who are not on your "publish" list. However, I
think that given other protocols (essentially, anything but XMPP) exist where
joining a chat will always expose your presence, I don't think it's reasonable
for clients or users to rely on this additional exposure not happening anyway.
Consider the absurdity of expecting that you can be invisible even when you've
joined a group chat. If someone wants the XMPP-specific semantics badly, we can
add a RoomPresence interface or somesuch, and they can set a property in order
to publish a per-room presence.

20080727032658-418b8-171a0e74e3910f331a49d333491376e1e73ec9f2.gz