Implement sha1_hex, sha1_bin in terms of GChecksum
Now that we use GLib >= 2.16, we can do this. I've kept the utility
wrappers because sha1_hex guarantees to return lower-case (unlike
GChecksum), and sha1_bin is more convenient than what's in GLib.

Also remove compatibility #define for G_PARAM_STATIC_STRINGS, because
we now depend on a GLib version that has it.