Remove the checks for the bytestream going from OPEN to INITIATING
A bug in the SOCSK5 code allowed the bytestream to emit a
connection-error signal when the stream was already OPEN, so the
fallback code had to check for the state going to OPEN back to
INITIATING (for a fallback stream).
The checks are not useful anymore now that the bug is fixed.