Fix auto-accepting MUC invites when changing presence
The patch to send our presence to the MUC when we go away/etc was broken
because it sent a <presence> stanza to MUC channels we'd been invited to
but not actually accepted yet, which automatically accepted the invites.
This was cool because Empathy didn't actually have a way to accept the
invite, but when I got back from lunch people were confused why I'd
joined the chat with Away / Lunch and then not said anything. :)