Don't let stream die before getting initial codecs and candidates
Constructing a GabbleMediaStream adds an idle callback to extract information
from its JingleContent only once the creator of the MediaStream has had a
chance to connect to its signals.  If the refcount of the MediaStream fell to
zero before that callback fired, it would crash.  So, the callback should ref
the stream.