fd.o#20003: GabbleJingle*: use GPOINTER_TO_SIZE for GTypes
This fixes a crash when the GType is instantiated on certain 64-bit
platforms. GType is defined to be the same size as gsize, which may
be greater than the size of int. On some 64-bit platforms (including
amd64 Linux?) the high bits of a GType are in practice zero anyway, so
the truncation is wrong but harmless and we didn't see this crash; on
other 64-bit platforms, we crash.