Silence the generation of gabble-signals-marshal.list
This rule fires every time, because if no source files have changed then
gabble-signals-marshal.list doesn't change so Make still thinks it's
old. The alternatives are to move the "build to temporary file and only
mv if it's changed" thing to building gabble-signals-marshal.h (which
would then run every time instead), or to touch the list even when it
hasn't changed (which would cause the whole of gabble to be rebuilt
whenever you change any source file).

I assert that this is better than rebuilding the whole of gabble. :-)