Create initial streams at once, not in an idle.
While it's true that by doing this from an idle they appear on the bus
after the media channel has been announced, that's not actually very
helpful: it doesn't mean the clients have had time to connect to the
relevant signals. In any case, the client should be calling ListStreams
on the media channel.