Don't unset updating_remote_codecs when s-e responds
This would fail in the case where we get a second description-info
before the streaming implementation has reacted to the first; if the
second one is deemed distasteful, Error() would be called and then the
stream would have been terminated, even though we want it to continue as
per XEP-0167 §10.

Also, remove FIXMEs about ack/nak-ing description-info: my reading of
the XEP says we don't need to.