Update (not replace) remote codecs on transport-info
The XEP says that clients SHOULD only send information about codecs
whose parameters have changed in transport-info. So, we have to update
Gabble's cache of remote codecs rather than just replacing it.

This patch also checks whether the name associated with an id has
changed, and if so NAKs the iq. If the remote end has sent information
for a new payload-type, it's just ignored.

The test had to be modified, because it was buggy:
 - it changed the codec names;
 - it sent all the codecs in the transport-info, rather than just the
   changes.