Delete vcard/test-vcard-set-and-get.py
It doesn't test anything useful that's not tested elsewhere.

It tests that RequestAvatar waits for the vCard to be retrieved before
returning or erroring, but that's already tested in test-avatar.py and
test-vcard-cache.py respectively.

It also somewhat tests that making two updates simultaneously works,
but that's tested in test-vcard-race.py (which doesn't actually test a
race! :D).