Only send <hold/> if the peer thinks they're unheld
We send <hold/> immediately we move to PendingHold, because we don't
expect holding to fail. We send <unhold/> only when we actually get back
to Unheld, because reacquiring the audio device etc. might fail.

If RequestHold(False) is called, and then before tpfs reacquires the
devices RequestHold(True) is called, we haven't yet informed the other
end that they're unheld, so we shouldn't send another <hold/>.