Remove most group flags from media channels
Removing the peer (to rescind the invitation or end the call) is wrong;
you should be closing the channel or removing yourself. As a result,
almost all the flag changes can go away, with the exception of having
CanAdd around until AddMembers/RequestStreams is called if a client is
using the request-anonymous-channel calling pattern.