Respond to removal of last content with terminate
XEP-0166 says (in a footnote!) that if your peer sends a content-remove
that leaves the session empty, you should send a session-terminate.

Making this work is a matter of deleting some code guarding against it
happening, and then deleting some other code (which never ran because of
the aforementioned guard) which was intended to implement it, but in
fact was redundant and crashed if allowed to run because the same "no
more contents ⇒ terminate" logic is also in content_removed_cb (which is
responsible for terminating the session rather than sending a
content-remove for the last content).

Deleting code is good, right?