Make GetAliases etc fall back to the full JID
Previously, Gabble fell back to the local part of the JID; if the
contact is foo@bar.com, it would default their alias to "foo". But this
is in violation of the spec, which says CMs should "[fall] back on the
handle name if [no alias] is present".

Gabble's old behaviour was slightly nicer on UIs that just want to show
something to the user, but make it impossible for UIs to tell whether a
contact actually has an alias or not. The real fix here is to separate
aliases into petnames and nicknames, but for now this makes Gabble
spec-compliant.