React to SupportedCodecs once when necessary.
For streams that we add, we never need to do anything in response to
SupportedCodecs. For streams that the peer adds, we need to send an
intersection exactly once.

Previously, the latter was mis-implemented as "ignore calls to
SupportedCodecs once the peer has sent us description-info". This worked
fine except when the peer sent description-info before we sent our
intersection.