Don't ignore subscription='remove' if remote-pending
In response to subscription='none' or subscription='from', the contact
is only removed from 'subscribe' if they're not remote pending., with
the explanation 'presence="unsubscribed" does that'. From my reading of
the RFC, that makes sense.

But I think the same guard for subscription='remove' makes no sense at
all: they're being removed entirely from the roster, so clearly we're
cancelling the subscription request.