Emit capabilities-discovered for the correct handle.
Previously, suppose contacts x and y have the same caps hash, and we
discoed x. When x replied, we'd satisfy both waiters, but emit
capabilities-discovered(x) for both. This should have made
RequestStreams erroneously never terminate in some cases (it's waiting
for caps that apparently never arrive) were it not for a bug in its
capabilities-discovered callback ...