Refactor satisfy/redisco/keep-waiting loop
I believe this logic to be approximately the same. I changed the "enough
trust" comparison not to subtract trust_inc. I believe that that was
subtracted to compensate for waiter_self being included in
disco_waiter_list_get_request_count (), but that's not necessary having
already removed waiter_self from waiters. (In fact it was not correct
before: waiter_self might already have been removed from waiters,
depending on the list's order.)