Don't assume session is non-NULL in GetCallState
If GetCallState() is called before RequestStreams() and friends,
there'll be no JingleSession for the channel, so we need to check if it
exists yet before dereferencing it.

It's valid to omit participants whose call state is 0, and if there's no
session we might not know who the peer is yet anyway (if the
RequestChannel(0) calling convention's in use).

Reviewed-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>