Unrace jingle/incoming-basics
We should expect the alleged gtalk4-compatibility transport-accept in
response to Gabble receiving a session-initiate, so move it up to the
first batch of expectations.

Also, picking up the initial StreamDirectionChanged is racy and not
generally possible in real clients (because the channel's only just been
announced, and might not have been dispatched to them by the time the
signal's emitted) so use ListStreams() instead.