Send <hold/> if necessary after session-initiate
If a call is put on hold before it's been initiated, Gabble should send
a <hold/> notification when the session-initiate is acked. This patch
makes this work both in the case where the GabbleJingleSession already
exists (but isn't initiated) when the call is placed on hold, and when
the session hasn't even been created yet (by setting the current hold
state on the session when it is created).