SOCKS5: decrease the connect timeout
When users are not on the same network or using IPv6, the direct
connection will always fail. Waiting 30 seconds before connecting to a
proxy is really long and is perceived by user as a problem in the FT or
stream tube. 10 seconds should be more than enough; if the connection
took more that 10 sec to establish, you'll probably have trouble to use
it any way.