Add a keepalive-interval connection parameter.
This allows clients to specify the interval between TCP and/or
whitespace keepalives (at the whim of Loudmouth) rather than hardcoding
the interval to 30 seconds (which is still the default). This might be
useful in situations where reducing network activity is more important
than realising the network has gone away quickly.