Ignore port=5222 if we think we know better.
Previously, if the UI sets port=5222 then Gabble disabled SRV lookups.
This turns out not to be great, because if the UI sets port=5222 (which
was not the default per Gabble's parameters, it is the spiritual XMPP
default port) then it can't connect to any XMPP server that uses an SRV
record.

While pedantically correct, the previous semantics turn out not to be
very helpful. We could go around fixing every account UI that ever gets
written, but ignoring 5222 if neither the connect server nor old-ssl
were set is more useful behaviour.