Implement C_Media_Capabilities_Immutable_Streams
The tests needed changing because the definition of the
Immutable_Streams cap implies that contacts who can only make audio
calls should have that flag set. Most of the tests were changed to make
the contacts video capable too; the jingle-caps test checks that Google
Video contacts (who are capable of both audio and video) have this flag
set.

This is a version of 744b0d3, modified to work with the refactored
capabilities code in 0.9