Expect ice-udp candidates in session-initiate.
The test previously expected an empty <transport/> stanza in the
session-initiate, followed by a transport-info action containing the
candidate. This matches gtalk-p2p, and is permitted by XEP-0176, but the
latter says the session-initiate SHOULD contain candidates.

So, test that the session-initiate contains the candidate we gave
Gabble, and also check that Gabble doesn't send a bonus empty
transport-info.