Update Jingle{Content,Session} to new Transport API
This ensures that inject_candidates and send_candidates are called at
all the appropriate moments, to give the transports the opportunities to
do what they need to do. Those appropriate moments are:

inject_candidates: when sending session-initiate, session-accept,
content-add, content-accept.

send_candidates: just after any of the above actions (for transports
that never include candidates in the actions and always send
transport-info), and when candidates are added for contents which have
already been announced.