Set LM_DEBUG=net and override the log handler when LM_DEBUG is not set.
This way, LM debug messages will be exposed through the Debug
interface when LM_DEBUG is not set (which is the majority of the
time). If LM_DEBUG is set, then it's a fair assumption that gabble is
being started on the command line, and therefore the Debug interface
isn't so important.

LM have some weird specific way of printing out their debugging
messages, so when LM_DEBUG has been set, I think it's appropriate not
to touch this default handler. Ideally, there would be a way to have
multiple handlers for a domain/level.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>