Test Google roster workarounds more explicitly
When I looked at this test, it wasn't clear to me that it actually
tested what it was supposed to test. Having looked more closely, I think
it was doing it right, but I think being more explicit about the
sequence of events we expect (and the events we do not want) will make
it easier for future readers.