Don't fake sub="remove" for "odd" google roster contacts
This was defeating the code that was supposed to work around a Google
Talk server bug whereby contacts "flicker" when moving to
ask="subscribe" and when moving from ask="subscribe" to
subscription="to/both". When a contact is actually removed, we assume
the server really meant it; this means you can still remove contacts who
are remote-pending by removing them from stored.

This still means that when you rescind a subscription request, if the
peer doesn't ack the <presence type="unsubscribe"/> with a <presence
type="unsubscribed"/> then the contact will stay on
subscribe:remote-pending until you reconnect.