Add some hysteresis to the Google Talk roster workaround.
We want to respect subscription requests being rescinded, but we need to
wait and see if they're actually being rescinded or whether Google's
lying to us. So if we receive a rescind notification just after an
ask=subscribe notification, we wait a second to see if Google's lying,
and only then actually remove them.