Don't hide contacts with gr:alias-for
Jonny Lamb has two JIDs on his Google account: his real JID, and his
gmail address. Google says that the first is an alias for the second,
but I'm actually subscribed to the first (and it's the one he uses).

I don't think hiding roster items with alias-for set is necessary.  We
hide Google roster items with subscription="none", which should catch
roster items which aren't real JIDs.