gabble_presence_cache_update_cache_entry: allow in == NULL
When the meaning of a caps bundle is forcibly inserted into the cache
with gabble_presence_cache_add_bundle_caps, it doesn't come with any
channel-manager-specific caps (the hash table in the struct describing
the bundle is NULL).

This could cause a crash when gabble_presence_cache_update_cache_entry is
called by _process_caps_uri, with the assumption that the hash table
is non-NULL.

This special-casing is currently done for one bundle, GTalk's voice-v1,
to work around failure to respond to disco correctly in old versions
of GTalk.

Also add a regression test, which models a real-world situation in which
this crash can occur - when a sipphone.com user tries to call an
@gmail.com JID (but apparently not an @googlemail.com JID!) as if it was
a SIP URI, sipphone.com's servers automatically bridge the call through
gtalk2voip.com, which makes the call by emulating an old version of GTalk.

Reviewed-by: Will Thompson <will.thompson@collabora.co.uk>