Add G_PARAM_CONSTRUCT to the fallback-stun-server property
(current) telepathy-glib adds default parameters in the parsed_params
structure, but doesn't set it in params_present. This causes the gabble
ConnectionManager implementation not to set default parameters explicitly on
the Connection. All other parameters work because they already have
G_PARAM_CONSTRUCT as a flag and have the default connection parameter as the
default for the property.