test-handles: stop exercising handle-holding by clients
This is telepathy-glib functionality, not Gabble functionality, so there
should be no need for us to test it here (I've copied this test into
telepathy-glib and adapted it, instead).

This test would also cause a critical warning in a future telepathy-glib,
because the bus name we're using is not actually syntactically valid.

Reviewed-by: Dafydd Harries <dafydd.harries@collabora.co.uk>