Fix disconnect-before-Google-relay-reply test
We do want to check that the Closed signal arrives for the channel when
you call Disconnect, because it doesn't in some cases. We also want to
check that the channel object actually goes away when you disconnect,
because it doesn't in some cases. This smokes out a crash in Gabble, but
sadly isn't reliable: libsoup seems to sometimes call HTTP request callbacks
as soon as you destroy the session, but other times uses an idle, and we
only crash in the latter case.