Don't ref media channel when waiting for relay
When we look up relay information from Google for new media streams, we
use a StreamCreationData * as the user_data for the libsoup HTTP request.
You can't cancel those requests, so the StreamCreationData can outlive the
call.

But currently, it holds a reference to the channel. This means that the
media channel can outlive the connection, which is bad for several
reasons:

 * It's still on the bus, so if you make a new connection and it happens
   to get the same path as the old one, then make a new call which happens
   to get the same path as this channel, we assert.
 * When the request does come back (or fails because we destroyed the
   SoupSession — this happens asynchronously), the channel's disposed. Its
   dispose() tries to call methods on the GabbleConnection, which is no
   longer valid, and we crash.

Instead, StreamCreationData now takes a weak ref to the channel. If the
channel's disposed, that field of the closure is set to NULL; the relevant
callbacks react to 'channel' being NULL by doing nothing, just as they do
if 'content' is NULL.