gabble_jingle_content_parse_add: clarify behaviour of libjingle 0.3 mode
When google_mode is TRUE and trans_node is NULL (i.e. the remote party
is using libjingle 0.3 or an emulation of it), the transport is implicitly
gtalk-p2p, because libjingle 0.3's protocol has no concept of transports.
To select this transport, we look up the empty string as a
pseudo-namespace in the factory's registry of transport types, which
in practice returns GABBLE_TYPE_JINGLE_TRANSPORT_GOOGLE.

This results in a coverity false positive a little further down,
where we select a transport based on trans_node (assumed to be non-NULL)
 if transport_type is zero. This is in fact safe, but coverity can't tell
that (it can't see that transport_type will be nonzero as a result of the
lookup), and it's not immediately obvious to programmers either.

To make this clearer, add a check that will, in practice, always succeed
(unless we delete the gtalk-p2p code for some strange reason).

Reviewed-by: Will Thompson <will.thompson@collabora.co.uk>