gabble_connection_set_self_capabilities: preserve all "legacy" capabilities
This is a stopgap solution to a misfeature of the SetSelfCapabilities
implementation: it assumes that "old-style" Capabilities are preserved.
Previously, this was arranged by not having any unified storage for
draft1 and legacy caps. However, that assumption is about to break.

So, the following policy is now used, which (perhaps by implementation
accident) matches the old behaviour: the new caps are the union of
(the old caps minus all "dynamic" caps) and the caps from channel
managers.

In practice "dynamic caps" means the caps for specific Tube services.