Separate capabilities into four sources (fixed, legacy Capabilities, and ContactCapabilities draft 1) and union them
This removes the need for the strange workaround in which
SetSelfCapabilities has to avoid deleting things that came from another
source.

To preserve historical behaviour (we support ice-udp and gtalk-p2p on
startup, until the first call to AdvertiseCapabilities), demote these to
optional, but put them in the initial value of the legacy
Capabilities set.