Split out the bonus caps seen at startup into a separate set; clear it when we're told our real caps
I can only assume that the rationale for having gtalk-p2p and ice-udp in
our initial presence is a hackish way to be callable as soon as we sign
in. Otherwise, there's a race between a UI making a VoIP call, and MC
pushing in our real caps; if the VoIP call happens too soon, the recipient
of the call will think we can only do ice-udp.

(This can be fixed better in MC, later.)