Bump raw-udp and ice-udp namespaces to :1
We never supported the <trying/> element, whose removal prompted the
namespace bump in raw-udp. The namespace bump in ice-udp is due to the
id='' attribute (which we support) and to keep the numbers the same
across Jingle specs.

There's no particularly pressing reason to support the old version: we
prefer to use gtalk-p2p if possible anyway, so we're
backwards-compatible with older Gabbles.

Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>