Recognize clients as media capable if they have any supported transport
Before gabble would only regard a client as audio and/or video capable when it
support Google p2p as a transport. For jingle clients it's ofcouse entirely
valid to not support Google p2p.

Allow clients to be marked as media capable if they support any of our
supported protocol (they still need to also advertise jingle/gtalk capabilities
seperatly. This is correct as we support all our transports as part of a normal
jingle session and for the google protocol it's not necessary to advertise a
transport.