test-avatar-multiple-resources.py: Don't expect AvatarUpdated for self_handle before conflict resolution
The D-Bus signal AvatarUpdated is usually emitted as soon as we receive a
presence stanza from a contact: we don't want to wait to receive the vCard with
the actual avatar because we don't ask the vCard unless an application
specifically request it. However, when we receive a presence stanza from
'ourselves' on another resource, we emit the signal AvatarUpdated only when the
conflict resolution is finished. This commit change the test to reflect that.