Copy FT per channel manager caps in their entirety.
In 0.8, FT caps are represented by two hash tables, mapping service
names to either NULL, or a Feature struct, depending on whether the caps
are for another user or for yourself.

The copy function assumed that we'd only ever copy other people's caps,
not our own, and so always set the value to NULL in the copy. But for
caching our own caps, and serving up disco responses from the cache, we
do want to copy the tables in their entirity.