Use (part of) D-Bus machine ID as default resource.
Previously, Gabble generated a random resource for each connection if no
resource was explicitly specified. This meant that you could connect to
the same account with two Telepathy-powered clients at the same time
(say, your N900 and Empathy on the desktop) without having to change
your resource on one of them.

However, this is too random: if you lose your network connection without
telling the server you're leaving, and then reconnect with a new
resource, the server may still think the old resource is connected
(until it decides it's timed out) and send messages there rather than to
your new connection. Ditto your contacts, if they're sending to full
JIDs. By using the D-Bus machine ID as the default resource, you'll get
the same resource each time you connect from a particular device, with a
reasonably good chance that it'll be different on each of your devices.

This has the downside that you can't make two connections to the same
account on the same bus. I don't think this is a major problem: MC
should be managing your connections anyway. :-) However, it does mean I
had to remove the test which checks that you can do that...