Turn the plugin loader into a fully-fledged object
GabblePluginLoader is a singleton, because it doesn't make sense to have
more than one plugin loader. It keeps a reference to itself once it's
instantiated, because it doesn't make sense to destroy it (our modules
can't be unloaded).

If Gabble is built without plugin support, GabblePluginLoader is still
an immortal singleton, it just doesn't do anything.