Don't use self if it is NULL.
Upon discussion with Sjoerd I learned that it was possible for a GAsync
function to be cancelled, but still return a valid result. In this case, that
would mean the GabbleJingleFactory was disposed. With my previous assumption
that the error handler would have caught any condition where the factory was
NULL, this could have tried to access the disposed factory. This commit fixes
that assumption.