fd.o #25296: break race in test-send-file-provide-immediately.py and test-send-file-wait-to-provide.py
bytestream.wait_bytestream_closed() may involve waiting for a socket
disconnection (if using SOCKS5). Waiting for the TransferredBytesChanged
before calling wait_bytestream_closed is racy as the ordering of socket and
D-Bus events is not garanteed.
We break this race by waiting for those signals *after* we called
wait_bytestream_closed(). As the TransferredBytesChanged signals are
manually catched using connect_to_signal we'll only wait for them if needed.