Make --disable-assumed-ft-cap work.
configure looks for args of the form "--enable-foo-bar" or
"--disable-foo-bar", and if if foo-bar is in a big table built from all
the AC_ARG_ENABLEs in configure.ac, it sets the variable enable_foo_bar.

So then our code would run, and set enable_assumed_ft_cap to yes,
trampling over any value it was set to by parsing the arguments to
configure.

The fourth argument of AC_ARG_ENABLE is the correct place to specify
what to do if the user doesn't pass anything about this option to
configure, so I moved the defaulting there; I also renamed the variable
to avoid the collision, consistent with the Werror arg.