Test that the tests waited for the stream-iq reply
By the time a test calls open_bytestream it should have waited for the response
to the stream-iq. If it doesn't races can occur which are hard to detect and
debug. So add a small boobytrap in open_bytestream which explodes if a test
didn't call check_si_reply, which it should have called after getting the reply