GabblePresenceCache: don't register the same LmMessageHandler twice
The presence cache wants to register the same C callback, with the same
user data, for two message types. The real Loudmouth allowed it to
use the same LmMessageHandler for each registration, but our
compatibility shim for Wocky doesn't support that usage; one of the
registrations just doesn't get called. In practice, this resulted in
the callback not being called for LM_MESSAGE_TYPE_MESSAGE.

(I've audited the rest of Gabble and this is the only place where this
happens.)