lm_connection_register_message_handler: remove bogus de-duplication
It is incorrect to register the same  twice: it won't work
with this implementation, and if done after connecting, it'll cause an
assertion failure. However, the current code silently ignores these
double-registrations if they happen before the connection is connected,
and omits the second and subsequent use from the list of delayed
registrations.

This patch causes both code paths to lead to the assertion failure,
which means we'll at least spot it if someone gets it wrong again :-)