Treat <item-not-found/> as the empty vCard
When we want to update our vCard, we first retrieve our vCard from the
server. XEP-0054 says that the server MUST return <item-not-found/> if
you have no vCard set, but we previously assumed that the server would
return success, with an empty <vCard/> node. Prosody respects the XEP.
So, let's treat the former case identically to the latter.

Fixes: fd.o#25987
Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>