Make sure the NewChannels signal contains the Call.
A race was happening where the testcase was failing due to the
NewChannels signal from the contact list being caught instead of
the one for the Call. This caused the rest of the test case to
continue before the Call was ready and would fail when trying
to retrieve the Contents property.