Add a variable with the path to Gabble to .pc files
Tests for out-of-tree Gabble plugins need to know where Gabble lives in
order to set its working directory and environment variables, etc. It's
possible to set the latter on the tests' dbus-daemon, but having the
path to Gabble allows plugins' test framework look more like Gabble's
(with their own directory of D-Bus services, rather than relying on the
installed set).

Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>