Add the MediaChannelRequest into pending_call_channels.
This looks like it was a typo. Another place in the code tries to
remove a MediaChannelRequest from pending_call_channels, but the
only place that inserts anything into pending_call_channels inserts
a channel. Inserting the MediaChannelRequest makes more sense as
then you can add request tokens to it.