Wait for capabilities before completely initializing a call.
If Gabble doesn't have a contact's capabilities, it should wait
until they are retrieved before returning from the CreateChannel
or EnsureChannel call. The caps must be checked to ensure they
have the required capabilities before returning whether or not
the call was successful.

Some of the capabilities checking code was copied from
media-channel.c.