Fix a non-UTF-8 string being sent to the debug log.
In the HTTP response for a Google relay session there is a
'magic_cookie' value which contains a binary 'magic' string. This
string, contained in the entire response, was being sent to the
debug log. If Empathy's debug window was open, this would cause
that string to be sent over DBus. Since the string wasn't valid
UTF-8, DBus dropped Gabble from the session bus. This patch escapes
the binary data in that output.