Fix tests with --enable-is-a-phone
Not only did two different C source files hardcode that Gabble is a PC,
but so did the test suite. This is why the tests passed when I only
changed one of the hardcoded C strings; I don't know why I thought they
passed when I fixed that, but... they didn't. :)