GabblePresenceCache: re-order signal emission vs. removal of pending requests
This fixes the following situation, which is difficult to test in the
regression tests without making them take even longer:

* the global "unsure period" for the presence cache has finished
  (5 seconds)
* we call a peer whose presence we do not have
* we ask that peer to decloak
* they do so, and give us a capabilities hash we don't already know about
* we send them a disco request to work out what it means
* they reply to the disco request

Previously, we'd signal to the media channel that it should proceed with
the delayed request, but because the node (caps hash) was still in our
table of pending disco requests, when the media channel asked whether
the presence cache was sure about the contact's capabilities yet, the
presence cache would reply that it was not, because it was waiting for
a disco request.

In the regression tests, this was masked by the fact that the
global "unsure period" had not finished. When it does, the media channel
sees this as another opportunity to retry pending calls, and by then,
the presence cache *does* consider itself to be sure about that contact.