ContactInfo: add a lookup table for known vCard fields
This contains everything that the current code hard-codes support for,
except KEY for which I don't think we have an adequate Telepathy
representation (it contains a Base64 blob, and exposing that on D-Bus
seems wrong).

Also, cache that information in hash tables keyed by the upper-case and
lower-case field names, as used in XMPP and Telepathy respectively.