_insert_edit_info: ignore type-parameters not starting with type=
The terminology is confusing. A vCard field (e.g. ADR) is called a "type"
in the jargon of RFC 2426; types have type parameters, which are key-value
pairs. The only key that has a mapping into XMPP is TYPE, which is defined
for several types including ADR; keys that we don't support include
LANGUAGE, ENCODING and VALUE.