Revert patch 2de68bb15afc8d1 and check for the dialect before setting the component name Simon's review of that patch had this explanation : I'm glad I caught this, because it's (subtly!) wrong. GabbleJingleTransport is the gtalk-p2p transport (used as an alternative to ICE-UDP or raw UDP), whereas the dialect you removed a check for indicates which dialect of Jingle we're using. (The choice is currently between XEP-0166 draft 0.32, XEP-0166 draft 0.15, the Jingle variant from libjingle 0.4, or the Jingle variant from libjingle 0.3 - at some point we should add XEP-0166 v1.0, which is not 100% compatible with 0.32, to that list.)
The official Google Talk clients will only ever use gtalk-p2p combined with
some Google Jingle variant, but Gabble is quite capable of using standard
XEP-0166 for the Jingle session, but the non-standard gtalk-p2p transport.
Indeed, until version 0.7.something we didn't support either of the standard
transports, and we still prefer to use gtalk-p2p if the target supports both
gtalk-p2p and ICE-UDP (because that lets us use Google's relays, if we're
signed in with a GMail account).