Pass the jingle factory a full jid instead of the resource
With Muji coming up and the ability to call bare jids on SIP gateways we can
call full jids from mucs, full jids for contacts and bare jids. This means that
juggling around the resource gets very error-prone. Luckily all the code that
wants to start a session already knows which jid it's for, so let them do the
hard work and have the API's just take jids directly