Removed broken dirty check before sending signal
The dirty check was broken because the mails_removed array is NULL
terminated so it's size is always bigger then one. Fixing this bug would
bring another one since when 30 emails is reached the count may changed
while the list of e-mails stays the same. For sake of simplicity, we simply
remove the check.