Make BaseChannel:requested writeable, in theory.
Most channels implement Requested as (initiator == conn->self_handle),
so that's how BaseChannel implements it. But if someone wants to
override it, then it has to be writeable. So: we make it writeable on
the base class, and make the default set implementation do nothing
except raise a warning if it's set to TRUE. (We can't warn if it's ever
set, because CONSTRUCT_ONLY properties are helpfully set to their
default values by GObject.)