Don't unblock contacts if they're removed from stored.
We want to ensure that clients which don't understand the 'deny' list
can remove contacts from other lists without accidentally unblocking
them. The only problematic case is 'stored': currently, the XMPP-level
response to removing someone from 'stored' is sending a roster update
removing them from the roster entirely. Since their blocked-ness is
stored on the roster, ...

So instead, when we would like to remove a blocked contact from the
roster, we instead cancel inbound and/or outbound subscriptions. This
has the effect of the server changing their subscription to 'none', at
which point they disappear from 'stored' (because on Google Talk we
hide subscription='none', ask≠'subscribe' contacts from stored).

There are two code paths for editing the roster, depending on whether we
have an edit in flight or not. The latter case is reasonably
straightforward. For the former case, we can be applying multiple edits
at once, so this patch moves the modification of the Google Roster state
to before the modification of the subscription state, to ensure that
queued modifications to the contact's blocked-ness have been processed
before we consider whether to handle removal specially.

Fixes fd.o#20597 (Blocked contacts get removed if they are removed from
stored)