Correctly default contacts on 'subscribe' to offline.
If we haven't received presence for a contact, we determine whether to
report them as offline or unknown based on our roster. In theory:

• If we've received the roster, and it shows that we should be receiving
  their presence, then we assume they're offline.
• Otherwise, we show them as unknown.

Unfortunately the existing logic is backwards: it shows contacts with
subscription='from' as offline, but should in fact be looking for
subscription='to'. This patch fixes this, and fixes the test.

Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>