Sanity-check conversion from gssize to socklen_t.
Coverity got upset that we were passing a signed variable as the third
argument of connect(), which is socklen_t, which is unsigned.

g_socket_address_get_native_size() only fails if gaddr is invalid, or if
the actual implementations (UnixSocketAddress/InetSocketAddress) are
internally inconsistent.

And sockaddr_storage is defined to be big enough to hold every supported
protocol-specific address structure, so...