Make XEP-0126 test more pedantic.
expect_many() doesn't enforce any ordering on the events, so the test
wasn't checking that commands were sent to the server in the right
order.

Also, checking that we were sending undirected presences revealed that
we were actually sending <presence to=''/>, that is, to the empty JID.
It turned out that we could simplify the success path of becoming
(in)visible while fixing this bug.