WIP: refactor setting initial presence
This patch adds a private structure for conn-presence, which currently
just holds the invisibility method in use. It then refactors
conn_presence_set_initial_presence_async() to be implemented in terms of
toggle_presence_visibility_async(). Finally, it removes the fallback
between mechanisms, because if a server claims to support (say) XEP-0186
invisibility, it can be reasonably expected to work. And, given that we
have to deal adequately with becoming invisible failing during a
connection, not just at login, we may as well deal in the same way in
both cases.

This patch is WIP because it breaks invisible_fallbacks.py (given that
it deletes the fallbacks that patch checks), because we should also have
tests for setting ourselves to invisible failing once the connection is
up and running (checking that Gabble reverts you to either your previous
presence, or DND, as we wish), and because we need to look up existing
privacy lists to fix this case:

• Gabble #1 signs in
• Gabble #1 creates a list called 'invisible', and activates it.
• Gabble #2 also signs in
• Gabble #2 tries to create a list called 'invisible', but it fails
  because you can't replace privacy lists that are in use, so it gives
  up on invisibility.